requirements.txt
matplotlib==3.7.1
numpy==1.24.2
Pillow==9.5.0
scikit_image==0.19.3
scikit_learn==1.2.0
torch==1.10.1+cu111
torchvision==0.11.2+cu111
tqdm==4.65.0
